progetto/requirements.txt
flask
flask_mqtt
flask_login
google-cloud-firestore